Add files via upload
Final Storm master database files for Ag and MSM sites. Nutrient (N parameters, DOC, manta temp) and physical data (pH, DO) for both sites is on the 30 minute scale, Hobo data (water level, hobo temp) is on the 15 minute scale. For MSM rain is on a 15 minute scale, but for Ag it's measured in a range of scales: each time the tipping bucket is activated (.01 inches of rain) it is recorded in date/time. This may result in Ag rain having multiple recordings within the same minute, or longer than a 30 minute scale depending on rainfall intensity. Graphing of the storms should understand these different date/time frequencies and adjust code accordingly.